tests: update imports and move metadata tests
Rework unit tests to follow recent refactors: move build_io_provenance_dict tests into a new core/test_metadata.py, adjust imports and symbol names (e.g. use io alias, resolve_output_path_from_metadata), and point overwrite-confirm mocks at napari_deeplabcut.ui.dialogs. Update test code paths that call the save dialog to use controls._save_layers_dialog. Remove/trim several obsolete UI-related tests (conflict summary/overwrite text and summarize_keypoint_conflicts) and replace some direct io monkeypatches with commented notes where mocking now belongs to the dialogs module.
